requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
